Extract class for chating
