+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/adpaters/LocalListAdapter.kt
@@ -4,7 +4,7 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.amigotrip.android.ZoomOutPageTransformer
+import com.amigotrip.android.custom.ZoomOutPageTransformer
 import com.amigotrip.android.datas.Article
 import com.amigotrip.anroid.R
 import kotlinx.android.synthetic.main.row_locals.view.*
